Finish re organise files into credit card module and stcpay module
