+++ sdk/src/main/java/com/moyasar/android/sdk/core/data/PaymentSource.kt
@@ -1,0 +1,3 @@
+package com.moyasar.android.sdk.core.data
+
+interface PaymentSource
