+++ sdk/src/main/java/com/moyasar/android/sdk/core/exceptions/ApiException.kt
@@ -1,5 +1,5 @@
 package com.moyasar.android.sdk.core.exceptions
 
-import com.moyasar.android.sdk.data.models.ErrorResponse
+import com.moyasar.android.sdk.core.data.response.ErrorResponse
 
 class ApiException(val response: ErrorResponse) : MoyasarException()
