+++ sdk/src/main/java/com/moyasar/android/sdk/creditcard/presentation/model/FieldValidation.kt
@@ -1,0 +1,8 @@
+package com.moyasar.android.sdk.creditcard.presentation.model
+
+enum class FieldValidation {
+  Name,
+  Number,
+  Expiry,
+  Cvc
+}
