+++ sdk/src/main/java/com/moyasar/android/sdk/creditcard/presentation/model/RequestResultViewState.kt
@@ -1,0 +1,7 @@
+package com.moyasar.android.sdk.creditcard.presentation.model
+
+
+internal sealed class RequestResultViewState<out T>{
+  data class Success<out T>(val data: T) : RequestResultViewState<T>()
+  data class Failure(val e: Exception) : RequestResultViewState<Nothing>()
+}
