--- sdk/src/main/java/com/moyasar/android/sdk/data/models/CardPaymentSource.kt
@@ -1,15 +1,0 @@
-package com.moyasar.android.sdk.data.models
-
-import com.google.gson.annotations.SerializedName
-
-data class CardPaymentSource(
-    val name: String,
-    val number: String,
-    val month: String,
-    val year: String,
-    val cvc: String,
-    val manual: String?,
-    @SerializedName("save_card") val saveCard: String?,
-    val type: String = "creditcard",
-    val token: String? = null,
-) : PaymentSource
