--- sdk/src/main/java/com/moyasar/android/sdk/data/models/ErrorResponse.kt
@@ -1,7 +1,0 @@
-package com.moyasar.android.sdk.data.models
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-import java.io.Serializable
-@Parcelize
-data class ErrorResponse(val message: String?, val type: String?, val errors: Map<String, List<String>>?) : Parcelable, Serializable
