--- sdk/src/main/java/com/moyasar/android/sdk/data/models/PaymentRequest.kt
@@ -1,12 +1,0 @@
-package com.moyasar.android.sdk.data.models
-
-import com.google.gson.annotations.SerializedName
-
-data class PaymentRequest(
-  val amount: Int,
-  val currency: String,
-  val description: String,
-  @SerializedName("callback_url") val callbackUrl: String,
-  val source: PaymentSource,
-  val metadata: Map<String, Any?> = HashMap()
-)
