--- sdk/src/main/java/com/moyasar/android/sdk/data/models/PaymentSource.kt
@@ -1,3 +1,0 @@
-package com.moyasar.android.sdk.data.models
-
-interface PaymentSource
