--- sdk/src/main/java/com/moyasar/android/sdk/domain/usecases/CreatePaymentUseCase.kt
@@ -1,8 +1,0 @@
-package com.moyasar.android.sdk.domain.usecases
-
-import com.moyasar.android.sdk.data.models.PaymentRequest
-import com.moyasar.android.sdk.data.remote.PaymentService
-
-class CreatePaymentUseCase(private val paymentService: PaymentService) {
-  suspend operator fun invoke(request: PaymentRequest) = paymentService.create(request)
-}
