--- sdk/src/main/java/com/moyasar/android/sdk/presentation/model/AuthResultViewState.kt
@@ -1,11 +1,0 @@
-package com.moyasar.android.sdk.presentation.model
-
-
-sealed class AuthResultViewState {
-  data class Completed(val id: String, val status: String, val message: String) :
-    AuthResultViewState()
-
-  data class Failed(val error: String? = null) : AuthResultViewState()
-
-  data object Canceled : AuthResultViewState()
-}
