--- sdk/src/main/java/com/moyasar/android/sdk/presentation/model/FieldValidation.kt
@@ -1,8 +1,0 @@
-package com.moyasar.android.sdk.presentation.model
-
-enum class FieldValidation {
-  Name,
-  Number,
-  Expiry,
-  Cvc
-}
